app/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
attrs==25.3.0
click==8.2.0
Faker==37.3.0
fastapi[all]==0.115.12
fastapi-cache2==0.2.2
frozenlist==1.6.0
greenlet==3.2.2
h11==0.16.0
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
multidict==6.4.3
pendulum==3.1.0
propcache==0.3.1
psycopg2-binary==2.9.10
pydantic==2.11.4
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
redis
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uvicorn==0.34.2
yarl==1.20.0
aioredis
